norminette + correction de protos

+++ libft/ft_memchr.c
@@ -1,27 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_atoi.c                                          :+:      :+:    :+:   */
+/*   ft_memchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: croussea <croussea@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/18 16:54:26 by croussea          #+#    #+#             */
-/*   Updated: 2023/11/20 18:07:01 by croussea         ###   ########.fr       */
+/*   Created: 2024/01/15 11:47:24 by croussea          #+#    #+#             */
+/*   Updated: 2024/01/15 11:49:32 by croussea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <stddef.h>
 
-void	*ft_memchr(const char *str, int c, unsigned int len)
+void	*ft_memchr(const char *str, int c, size_t len)
 {
-    int i;
-    
-    i = 0;
+	size_t	i;
+
+	i = 0;
 	while (i < len && str[i] != c)
 	{
 		i++;
 	}
-    if (i < len && str[i] == c)
+	if (i < len && str[i] == c)
 		return ((void *)(unsigned char *)str + i);
 	else
 		return (NULL);
-}	
+}
