+++ libft/ft_memcpy.c
@@ -1,8 +1,20 @@
-void *ft_memcpy(void *dest, const void *src, unsigned int num)
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: croussea <croussea@student.42madrid>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/15 11:52:38 by croussea          #+#    #+#             */
+/*   Updated: 2024/01/15 11:53:33 by croussea         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+void	*ft_memcpy(void *dest, const void *src, size_t num)
 {
-	unsigned char *d;
-	const unsigned char *s;
-	unsigned int i;
+	unsigned char		*d;
+	const unsigned char	*s;
+	size_t      		i;
 
 	i = 0;
 	d = dest;
