+++ libft/ft_memset.c
@@ -1,7 +1,19 @@
-void *ft_memset(void *ptr, int value, unsigned int num)
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memset.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: croussea <croussea@student.42madrid>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/15 11:56:20 by croussea          #+#    #+#             */
+/*   Updated: 2024/01/15 11:56:45 by croussea         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+void	*ft_memset(void *ptr, int value, size_t num)
 {
-	unsigned char *p;
-	unsigned int i;
+	unsigned char	*p;
+	size_t	i;
 
 	p = ptr;
 	i = 0;
